+++ tests/covalent_tests/workflow/electron_test.py
@@ -305,7 +305,7 @@ def test_autogen_dict_electrons():
     assert set(g.edges) == {(1, 0, 0), (3, 1, 0), (2, 1, 0), (0, 4, 0)}
 
 
-def test_as_transportable_dict(mocker):
+def test_as_transportable_dict():
     """Test the get transportable electron function."""
 
     @ct.electron
@@ -323,7 +323,7 @@ def test_as_transportable_dict(mocker):
     assert TransportableObject(test_func).to_dict() == transportable_electron["function"]
 
 
-def test_call_sublattice(mocker):
+def test_call_sublattice():
     """Test the sublattice logic when the __call__ method is invoked."""
 
     @ct.lattice(executor="mock")
@@ -349,3 +349,138 @@ def test_call_sublattice(mocker):
                 assert (
                     node_data["function"].get_deserialized().__name__ == "_build_sublattice_graph"
                 )
+
+
+def test_electron_auto_task_groups():
+    @ct.electron
+    def task(arr: list):
+        return sum(arr)
+
+    @ct.electron
+    @ct.lattice
+    def sublattice(x):
+        return task(x)
+
+    @ct.lattice
+    def workflow(x):
+        return sublattice(x)
+
+    workflow.build_graph([[1, 2], 3])
+    tg = workflow.transport_graph
+    assert all(tg.get_node_value(i, "task_group_id") == 0 for i in [0, 3, 4])
+    assert all(tg.get_node_value(i, "task_group_id") == i for i in [1, 2, 5, 6, 7, 8])
+
+
+def test_electron_get_attr():
+    class Point:
+        def __init__(self, x, y):
+            self.x = x
+            self.y = y
+
+    @ct.electron
+    def create_point():
+        return Point(3, 4)
+
+    @ct.electron
+    def add(a, b):
+        return a + b
+
+    @ct.lattice
+    def workflow():
+        point = create_point()
+        return add(point.x, point.y)
+
+    workflow.build_graph()
+    tg = workflow.transport_graph
+
+    # TG:
+    # 0: point
+    # 1: point.__getattr__
+    # 2: "x"
+    # 3: point.__getattr__
+    # 4: "y"
+    # 5: add
+    # 6: "postprocess"
+
+    point_electron_gid = tg.get_node_value(0, "task_group_id")
+    getitem_x_gid = tg.get_node_value(1, "task_group_id")
+    getitem_y_gid = tg.get_node_value(3, "task_group_id")
+    assert point_electron_gid == 0
+    assert getitem_x_gid == point_electron_gid
+    assert getitem_y_gid == point_electron_gid
+    assert all(tg.get_node_value(i, "task_group_id") == i for i in [2, 4, 5, 6])
+
+
+def test_electron_auto_task_groups_getitem():
+    """Test task packing with __getitem__"""
+
+    @ct.electron
+    def create_array():
+        return [3, 4]
+
+    @ct.electron
+    def add(a, b):
+        return a + b
+
+    @ct.lattice
+    def workflow():
+        arr = create_array()
+        return add(arr[0], arr[1])
+
+    workflow.build_graph()
+    tg = workflow.transport_graph
+
+    # TG:
+    # 0: arr
+    # 1: arr.__getitem__
+    # 2: 0
+    # 3: arr.__getitem__
+    # 4: 1
+    # 5: add
+    # 6: "postprocess"
+
+    arr_electron_gid = tg.get_node_value(0, "task_group_id")
+    getitem_x_gid = tg.get_node_value(1, "task_group_id")
+    getitem_y_gid = tg.get_node_value(3, "task_group_id")
+    assert arr_electron_gid == 0
+    assert getitem_x_gid == arr_electron_gid
+    assert getitem_y_gid == arr_electron_gid
+    assert all(tg.get_node_value(i, "task_group_id") == i for i in [2, 4, 5, 6])
+
+
+def test_electron_auto_task_groups_iter():
+    """Test task packing with __iter__"""
+
+    @ct.electron
+    def create_tuple():
+        return (3, 4)
+
+    @ct.electron
+    def add(a, b):
+        return a + b
+
+    @ct.lattice
+    def workflow():
+        tup = create_tuple()
+        x, y = tup
+        return add(x, y)
+
+    workflow.build_graph()
+    tg = workflow.transport_graph
+
+    # TG:
+    # 0: tup
+    # 1: tup.__getitem__
+    # 2: 0
+    # 3: tup.__getitem__
+    # 4: 1
+    # 5: add
+    # 6: "postprocess"
+
+    tup_electron_gid = tg.get_node_value(0, "task_group_id")
+    getitem_x_gid = tg.get_node_value(1, "task_group_id")
+    getitem_y_gid = tg.get_node_value(3, "task_group_id")
+    assert tup_electron_gid == 0
+    assert getitem_x_gid == tup_electron_gid
+    assert getitem_y_gid == tup_electron_gid
+    assert all(tg.get_node_value(i, "task_group_id") == i for i in [2, 4, 5, 6])