+++ tests/covalent_tests/workflow/lattice_test.py
@@ -23,7 +23,7 @@
 from dataclasses import asdict
 
 import covalent as ct
-from covalent._shared_files.defaults import DefaultMetadataValues
+from covalent._shared_files.defaults import DefaultMetadataValues, postprocess_prefix
 
 DEFAULT_METADATA_VALUES = asdict(DefaultMetadataValues())
 
@@ -86,3 +86,24 @@ def test_lattice_executor_settings():
     assert workflow.metadata["executor"] == DEFAULT_METADATA_VALUES["executor"]
     workflow_2.build_graph(1)
     assert workflow_2.metadata["executor"] == "custom_executor"
+
+
+def test_lattice_build_graph(mocker):
+    """Test the build graph method in lattice."""
+
+    @ct.electron
+    def task(x):
+        return x
+
+    @ct.lattice
+    def workflow(x):
+        return task(x)
+
+    original_exhaustive_value = ct._shared_files.config.get_config("sdk.exhaustive_postprocess")
+    ct._shared_files.config.set_config("sdk.exhaustive_postprocess", "false")
+    workflow.build_graph(1)
+    assert workflow.transport_graph.get_node_value(2, "name") == f"{postprocess_prefix}reconstruct"
+    ct._shared_files.config.set_config("sdk.exhaustive_postprocess", "true")
+    workflow.build_graph(1)
+    assert workflow.transport_graph.get_node_value(2, "name") == postprocess_prefix
+    ct._shared_files.config.set_config("sdk.exhaustive_postprocess", original_exhaustive_value)